+++ experimental/instrumentation-replay/jest.config.js
@@ -1,0 +1,7 @@
+const { jestBaseConfig } = require('../../jest.config.base.js');
+
+module.exports = {
+  ...jestBaseConfig,
+  roots: ['experimental/instrumentation-replay/src'],
+  testEnvironment: 'jsdom',
+};